B up

+++ db/migrate/20210710144921_create_friendly_id_slugs.rb
@@ -10,8 +10,8 @@ MIGRATION_CLASS =
 class CreateFriendlyIdSlugs < MIGRATION_CLASS
   def change
     create_table :friendly_id_slugs do |t|
-      t.string   :slug,           null: false
-      t.integer  :sluggable_id,   null: false
+      t.string   :slug, null: false
+      t.bigint :sluggable_id, null: false
       t.string   :sluggable_type, limit: 50
       t.string   :scope
       t.datetime :created_at, precision: 6, default: -> { 'CURRENT_TIMESTAMP' }, null: false
