+++ db/migrate/20210712020439_create_pghero_query_stats.rb
@@ -6,9 +6,9 @@ class CreatePgheroQueryStats < ActiveRecord::Migration[6.1]
       t.text :database
       t.text :user
       t.text :query
-      t.integer :query_hash, limit: 8
+      t.bigint :query_hash, limit: 8
       t.float :total_time
-      t.integer :calls, limit: 8
+      t.bigint :calls, limit: 8
       t.timestamp :captured_at
     end
